Second Commit: Original Quantity added in item and modified the role permission as 2 roles

+++ Services/Implementations/ItemService.cs
@@ -60,8 +60,9 @@ public class ItemService : IItemService
         {
             ItemId = Ulid.NewUlid(),
             ItemName = itemDto.ItemName,
+            OriginalQuantity = itemDto.OriginalQuantity,
             Quantity = itemDto.Quantity,
-            Place = itemDto.Place,
+            Location = itemDto.Location,
             FileId = fileName!,
             Description = itemDto.Description,
             CreatedBy = string.IsNullOrEmpty(userEmail) ? Creator.System : userEmail,
@@ -71,6 +72,7 @@ public class ItemService : IItemService
 
         // save the metadata to db
         await _itemRepo.Create(newItem);
+        await _itemRepo.SaveChangesAsync();
         return new OperationResult(true, "successfully created", null);
     }
 
@@ -93,9 +95,9 @@ public class ItemService : IItemService
         return new OperationResult(true, "File deleted successfully",null);
     }
 
-    public async Task<PaginatedList<ItemDetails>> GetAll(string? keyWord)
+    public async Task<PaginatedList<ItemDetails>> GetAll(ItemQueryParams queryParams)
     {
-        var ItemList = await _itemRepo.GetAllDetails(keyWord);
+        var ItemList = await _itemRepo.GetAllDetails(queryParams);
         return ItemList;
     }
 
@@ -108,41 +110,51 @@ public class ItemService : IItemService
     {
         var itemInDb = await _itemRepo.GetById(itemId);
         if (itemInDb is null)
-            return new OperationResult(false, $"There is no item with id {itemId}",null);
+            return new OperationResult(false, $"There is no item with id {itemId}", null);
 
         var userEmail = _httpContextAccessor
             .HttpContext?.User?.FindFirst("email")?.Value;
         userEmail = string.IsNullOrEmpty(userEmail) ? Creator.System : userEmail;
 
         // try uploading
-        if (file is not null || file?.Length != 0)
+        try
         {
-            var createResult = await _fileService.Upload(file!);
-            if (!createResult.Success)
-                return new OperationResult(false, "Failed while file uploading, Please try again", null);
-            await _fileMetaDataRepo.Delete(itemInDb.FileId);
-            itemInDb.FileId = createResult.Name!;
-            var newFileMetaData = new FileMetadata
+            if (file is not null && file?.Length != 0)
             {
-                GeneratedName = file!.FileName!,
-                OriginalName = file!.FileName,
-                ContentType = file.ContentType,
-                UploadedBy = userEmail,
-            };
-            await _fileMetaDataRepo.Create(newFileMetaData);
-
+                var createResult = await _fileService.Upload(file!);
+                if (!createResult.Success)
+                    return new OperationResult(false, "Failed while file uploading, Please try again", null);
+                await _fileMetaDataRepo.Delete(itemInDb.FileId);
+                itemInDb.FileId = createResult.Name!;
+                var newFileMetaData = new FileMetadata
+                {
+                    GeneratedName = file!.FileName!,
+                    OriginalName = file!.FileName,
+                    ContentType = file.ContentType,
+                    UploadedBy = userEmail,
+                };
+                await _fileMetaDataRepo.Create(newFileMetaData);
+            }
+            Ulid.TryParse(itemDetails?.CategoryId, out var catId);
+            //save new metadata
+            if (itemDetails is not null && itemDetails.ModifiedAt == itemInDb.ModifiedAt)
+            {
+                itemInDb.ItemName = itemDetails.ItemName;
+                itemInDb.OriginalQuantity = itemDetails.OriginalQuantity;
+                itemInDb.Quantity = itemDetails.Quantity;
+                itemInDb.Location = itemDetails.Location;
+                itemInDb.Description = itemDetails.Description;
+                itemInDb.CategoryId = catId;
+                itemInDb.ModifiedAt = DateTimeOffset.UtcNow;
+                itemInDb.ModifiedBy = userEmail;
+                await _itemRepo.SaveChangesAsync();
+                return new OperationResult(true, "updated successfully", itemInDb.FileId);
+            }
+            return new OperationResult(false, "The record has been modified by another user. Please reload and try again.", null);
         }
-
-        //save new metadata
-        if (itemDetails is not null)
+        catch (Exception ex)
         {
-            itemInDb.ItemName = itemDetails.ItemName;
-            itemInDb.Quantity = itemDetails.Quantity;
-            itemInDb.Place = itemDetails.Place;
-            itemInDb.Description = itemDetails.Description;
-            itemInDb.CategoryId = itemInDb.CategoryId;
+            return new OperationResult(false, $"There was an error {ex.Message}", null);
         }
-        await _itemRepo.SaveChangesAsync();
-        return new OperationResult(true, "updated successfully", itemInDb.FileId);
     }
 }
